Fix: support large file sizes on 32 bit archs
Fixes the following error message:

```
cannot unmarshal number into Go struct field
ExifInfo.Assets.items.exifInfo.fileSizeInByte of type int
```

This is caused by a video in immich that is larger than can fit into an
int32.

+++ ui/size.go
@@ -5,7 +5,7 @@ import (
 	"math"
 )
 
-func FormatBytes(s int) string {
+func FormatBytes(s int64) string {
 	suffixes := []string{"B", "KB", "MB", "GB"}
 	bytes := float64(s)
 	base := 1024.0
